pyproject.toml
[project]
name = "trade-claw"
version = "0.1.0"
description = "Streamlit app for Kite Connect historical prices (HDFC Bank)"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.40.0",
    "python-dotenv",
    "streamlit",
    "kiteconnect",
    "pandas",
    "plotly",
    "mcp>=1.26.0",
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.42.0",
    "celery[redis]>=5.6.2",
    "redis>=6.4.0",
    "langgraph>=1.1.3",
    "langchain-openai>=1.1.11",
]
